Remove HnStoryList component and refactor ServerStoryList to utilize HnStoryListBody and HnStoryListToggle for improved structure and separation of concerns.
