--- src/features/storyList/HnStoryList.tsx
@@ -1,187 +1,0 @@
-import { createMemo, For, Show } from "solid-js";
-
-import { Shell } from "~/components/Icon";
-import { PullToRefresh } from "~/components/PullToRefresh";
-import {
-  useDataStore,
-  useReadItemsStore,
-  useRefreshStore,
-  useServiceWorkerStore,
-} from "~/contexts/AppDataContext";
-import { useSortFunction } from "~/hooks/useSortFunction";
-import { timeSince } from "~/lib/utils";
-
-import { HnListItem } from "./HnListItem";
-
-import type {
-  HnStorySummary,
-  StoryPage,
-  TopStoriesType,
-} from "~/models/interfaces";
-
-interface HnStoryListProps {
-  items?: HnStorySummary[];
-
-  page: TopStoriesType;
-}
-
-export function HnStoryList(props: HnStoryListProps) {
-  const dataStore = useDataStore();
-  const readItemsStore = useReadItemsStore();
-  const refreshStore = useRefreshStore();
-
-  const serviceWorker = useServiceWorkerStore();
-  const isOffline = serviceWorker.isOfflineMode;
-  const isLoading = () => dataStore.isLoadingData();
-  const onRefresh = () => dataStore.refreshActive();
-
-  // server will not have any stories to render since we rely on createEffect
-  // render 0 here to avoid hydration mismatch
-  const itemsToRender = () => {
-    const items = props.items;
-    if (!items) {
-      return undefined;
-    }
-    const sortType = props.page === "topstories" ? undefined : "score";
-    return useSortFunction(items, sortType) ?? [];
-  };
-
-  const lastUpdatedTs = createMemo(() => {
-    const page = props.page as StoryPage;
-    return refreshStore.refreshTimestamps[page];
-  });
-
-  const lastRequestedTs = createMemo(() => {
-    const page = props.page as StoryPage;
-    return refreshStore.refreshRequestedTimestamps[page];
-  });
-
-  const pullMessage = createMemo(() => {
-    return lastUpdatedTs()
-      ? `Updated ${timeSince(lastUpdatedTs(), true)}`
-      : undefined;
-  });
-
-  const requestMessage = createMemo(() => {
-    return lastRequestedTs()
-      ? `Requested ${timeSince(lastRequestedTs(), true)}`
-      : undefined;
-  });
-
-  const toggleHideReadItems = () => {
-    readItemsStore.setReadSettings(
-      "shouldHideReadItems",
-      !readItemsStore.readSettings.shouldHideReadItems
-    );
-  };
-
-  return (
-    <>
-      <Show
-        when={(itemsToRender() ?? []).length > 0}
-        fallback={
-          <div class="text-center text-lg text-gray-500">
-            No items to show. Most likely, you are filtering to hide read items.
-            Click the toggle above to change the setting.
-          </div>
-        }
-      >
-        <PullToRefresh
-          disabled={isLoading()}
-          onRefresh={onRefresh}
-          // message={pullMessage()}
-        >
-          <Show when={pullMessage() || requestMessage()}>
-            <div class="flex flex-wrap items-center justify-center gap-x-4 gap-y-1 text-center text-[11px] text-slate-400">
-              <button
-                type="button"
-                title="Refresh now"
-                onClick={() => {
-                  if (isOffline() || isLoading()) {
-                    return;
-                  }
-                  onRefresh();
-                }}
-                class="inline-flex items-center gap-1 hover:text-orange-500 focus:outline-none active:text-orange-500"
-                aria-label="Refresh list"
-              >
-                <span
-                  class={
-                    isLoading() ? "inline-flex animate-spin" : "inline-flex"
-                  }
-                >
-                  <Shell width="12" height="12" />
-                </span>
-                <span>{pullMessage() ?? "Refresh"}</span>
-              </button>
-              <Show when={requestMessage()}>
-                <button
-                  type="button"
-                  title="Refresh now"
-                  onClick={() => {
-                    if (isOffline() || isLoading()) {
-                      return;
-                    }
-                    onRefresh();
-                  }}
-                  class="text-[10px] text-slate-400 hover:text-orange-500 focus:outline-none active:text-orange-500"
-                  aria-label="Refresh list"
-                >
-                  {requestMessage()}
-                </button>
-              </Show>
-            </div>
-          </Show>
-          <div class="grid grid-cols-[1fr_1fr_1fr_3fr]">
-            <For each={itemsToRender() ?? []}>
-              {(item) => {
-                const isRead = () =>
-                  readItemsStore.readItems[item.id] !== undefined;
-                const isRecentRead = () =>
-                  readItemsStore.recentlyReadId() === item.id;
-                const shouldRender = () =>
-                  // allow SSR hydration match, or if not hiding read items
-                  !readItemsStore.readSettings.shouldHideReadItems ||
-                  !isRead() ||
-                  isRecentRead();
-
-                return (
-                  <Show when={shouldRender()}>
-                    <HnListItem
-                      data={item}
-                      recentFadeOut={
-                        isRecentRead() &&
-                        readItemsStore.readSettings.shouldHideReadItems
-                      }
-                      onFadeComplete={() =>
-                        readItemsStore.setRecentlyReadId(undefined)
-                      }
-                    />
-                  </Show>
-                );
-              }}
-            </For>
-          </div>
-        </PullToRefresh>
-      </Show>
-      <div class="flex items-center gap-2 pt-8">
-        <label class="inline-flex cursor-pointer items-center gap-2">
-          <input
-            type="checkbox"
-            checked={readItemsStore.readSettings.shouldHideReadItems}
-            onChange={toggleHideReadItems}
-            class="peer sr-only"
-          />
-
-          <div class="peer relative h-6 w-11 rounded-full bg-gray-200 peer-checked:bg-orange-600 peer-focus:ring-4 peer-focus:ring-orange-300 peer-focus:outline-none after:absolute after:start-[2px] after:top-[2px] after:h-5 after:w-5 after:rounded-full after:border after:border-gray-300 after:bg-white after:transition-all after:content-[''] peer-checked:after:translate-x-full peer-checked:after:border-white rtl:peer-checked:after:-translate-x-full dark:border-gray-600 dark:bg-gray-700 dark:peer-focus:ring-orange-800" />
-
-          <span>
-            {readItemsStore.readSettings.shouldHideReadItems
-              ? "Hiding read items (click to show)"
-              : "Showing read items (click to hide)"}
-          </span>
-        </label>
-      </div>
-    </>
-  );
-}
